Formatted al documents, added missing init and upload report models.
All documents were formatted, delegated labels dtype to data.py: BCE requires long but CE double. Removed torch version greater support. Removed legacy main.py

+++ src/utils/train_functions.py
@@ -44,7 +44,7 @@ def train_step(
         outputs = model(inputs)
 
         # Compute loss
-        loss_value = loss(outputs, targets.double())
+        loss_value = loss(outputs, targets)
         losses.append(loss_value.item())
 
         optimizer.zero_grad()
@@ -95,7 +95,7 @@ def val_step(
             outputs = model(inputs)
 
             # Compute loss
-            loss_value = loss(outputs, targets.double())
+            loss_value = loss(outputs, targets)
             losses.append(loss_value.item())
 
             # Update accuracy
